Nick's update

+++ website/views.py
@@ -14,14 +14,17 @@ def frontpage():
 @views.route('/', methods=['GET', 'POST'])
 @login_required
 def home():
-    if request.method == 'POST': 
-        note = request.form.get('note')#Gets the note from the HTML 
+    if request.method == 'POST':
+        if current_user.role != 'manager':
+            flash('You are not authorized to perform this action', category='error')
+            return redirect(url_for('views.home'))
 
+        note = request.form.get('note')
         if len(note) < 1:
-            flash('Note is too short!', category='error') 
+            flash('Note is too short!', category='error')
         else:
-            new_note = Note(data=note, user_id=current_user.id)  #providing the schema for the note 
-            db.session.add(new_note) #adding the note to the database 
+            new_note = Note(data=note, user_id=current_user.id)
+            db.session.add(new_note)
             db.session.commit()
             flash('Note added!', category='success')
 
@@ -29,13 +32,14 @@ def home():
 
 
 @views.route('/delete-note', methods=['POST'])
-def delete_note():  
-    note = json.loads(request.data) # this function expects a JSON from the INDEX.js file 
+@login_required
+def delete_note():
+    note = json.loads(request.data)
     noteId = note['noteId']
     note = Note.query.get(noteId)
-    if note:
-        if note.user_id == current_user.id:
-            db.session.delete(note)
-            db.session.commit()
-
+    if note and note.user_id == current_user.id:
+        if current_user.role != 'manager':
+            return jsonify({'error': 'Unauthorized'}), 403
+        db.session.delete(note)
+        db.session.commit()
     return jsonify({})
